requirement.txt
torch==1.1.0
torchvision==0.3
mmcv==0.2.12
editdistance
Polygon3
pyclipper
opencv-python==3.4.2.17
Cython